+++ src/redux/actions/filter-actions.js
@@ -1,0 +1,19 @@
+import { actionTypes } from '../constants';
+
+const { CHANGE_FILTER, SEARCH_ITEM } = actionTypes;
+
+const changeFilter = (activeFilter) => {
+  return {
+    type: CHANGE_FILTER,
+    payload: activeFilter,
+  };
+};
+
+const searchItem = (value) => {
+  return {
+    type: SEARCH_ITEM,
+    payload: value,
+  };
+};
+
+export { changeFilter, searchItem };
