+++ src/redux/constants.js
@@ -5,6 +5,8 @@ const actionTypes = {
   CHANGE_FILTER: 'CHANGE_FILTER',
   SEARCH_ITEM: 'SEARCH_ITEM',
   EDIT_TASK_TITLE: 'EDIT_TASK_TITLE',
+  SET_TASK_AUTHOR: 'SET_TASK_AUTHOR',
+  SAVE_TASK_ON_SERVER: 'SAVE_TASK_ON_SERVER',
   LOGIN: 'LOGIN',
   LOGOUT: 'LOGOUT',
   ADD_ROLE: 'ADD_ROLE',
